SIMD: Break out large header into smaller files in base/SIMD/
